CORE-7089 Distribute group parameters after approval
